projekat/requirements.txt
RPi.GPIO
paho-mqtt
adafruit-circuitpython-dht
rpi_lcd
